(WIP) Implementing proposal execution depending on proposal type

+++ libraries/chain/include/deip/chain/research_group_object.hpp
@@ -24,7 +24,7 @@ public:
     fc::shared_string permlink;
     fc::shared_string desciption;
 
-    u_int16_t quorum_percent;
+    uint32_t quorum_percent;
     share_type total_tokens_amount;
     flat_set<research_group_token_object> research_group_tokens;
 };
@@ -40,7 +40,6 @@ public:
 
 public:
     research_group_token_id_type id;
-
     research_group_id_type research_group;
     share_type amount;
     account_name_type owner;
